requirements.txt
fastapi
uvicorn
psycopg2
python-dotenv
face_recognition
opencv-python
dlib
